1. Move from the "TaskInTasksList.js" file to the "helper.js" file the logic to display the date and time of task creation in the local time zone; 2. Change the appearance of displaying the date and time of creating tasks in an open task.
